Refactor names and decoupled Classes
